@samsoares 2018-08-29
Can you show me how you got this perf improvements?